Install Auto-Correct in Termux
If you are a new termux User or you are just a fast type and do some spelling mistakes then this one is Specially for you. using this Command we will install an Auto-Correct In termux and it will give you an indication whether you are typing the right command or not.